Add the extra dependencies required for the project
To complete the project, requires PHP's International extension for
formatting values as a currency, Twilio's PHP SDK to interact with
Twilio's APIs, and Symfony's EventDispatcher, as that's the core focus
of what the application's demonstrating.

I added in the Monolog bundle to do some debugging during development
and the Maker bundle to simple creating Controllers (and potential other
classes).